Rename EvaluationMetric to ErrorMetric
