+++ include/albatross/GP
@@ -16,6 +16,7 @@
 #include "Core"
 #include "CovarianceFunctions"
 
+#include <albatross/evaluation/likelihood.h>
 #include <albatross/eigen/serializable_ldlt.h>
 #include <albatross/models/gp.h>
