+++ include/albatross/evaluation/evaluate.h
@@ -15,113 +15,7 @@
 
 namespace albatross {
 
-/*
- * An EvaluationMetric is basically just a wrapper around a
- * function which enforces a signature used in evaluation.  One
- * alternative would be to use std::function.  Ie,
- *
- *   template <typename RequiredPredictType>
- *   using EvaluationMetric =
- *         std::function<double(const RequiredPredictType &,
- *                              const MarginalDistribution &);
- *
- * the problem with this is that when you pass the std::function
- * object as an argument the RequiredPredictType cannot be
- * inferred.
- */
 
-template <typename RequiredPredictType>
-using Evaluator = double (*)(const RequiredPredictType &,
-                             const MarginalDistribution &);
-
-template <typename RequiredPredictType> struct EvaluationMetric {
-
-  Evaluator<RequiredPredictType> eval_;
-
-  EvaluationMetric(Evaluator<RequiredPredictType> eval) : eval_(eval) {}
-
-  double operator()(const RequiredPredictType &prediction,
-                    const MarginalDistribution &truth) const {
-    return eval_(prediction, truth);
-  }
-
-  template <typename ModelType, typename FeatureType, typename FitType>
-  double
-  operator()(const Prediction<ModelType, FeatureType, FitType> &prediction,
-             const MarginalDistribution &truth) const {
-    return (*this)(prediction.template get<RequiredPredictType>(), truth);
-  }
-
-  template <class Archive> void serialize(Archive &){};
-};
-
-static inline double root_mean_square_error(const Eigen::VectorXd &prediction,
-                                            const Eigen::VectorXd &truth) {
-  const Eigen::VectorXd error = prediction - truth;
-  double mse = error.dot(error) / static_cast<double>(error.size());
-  return sqrt(mse);
-}
-
-static inline double root_mean_square_error(const Eigen::VectorXd &prediction,
-                                            const MarginalDistribution &truth) {
-  return root_mean_square_error(prediction, truth.mean);
-}
-
-struct RootMeanSquareError : public EvaluationMetric<Eigen::VectorXd> {
-  RootMeanSquareError()
-      : EvaluationMetric<Eigen::VectorXd>(root_mean_square_error) {}
-};
-
-static inline double standard_deviation(const Eigen::VectorXd &prediction,
-                                        const Eigen::VectorXd &truth) {
-  Eigen::VectorXd error = prediction - truth;
-  const auto n_elements = static_cast<double>(error.size());
-  const double mean_error = error.sum() / n_elements;
-  error.array() -= mean_error;
-  return std::sqrt(error.dot(error) / (n_elements - 1));
-}
-
-static inline double standard_deviation(const Eigen::VectorXd &prediction,
-                                        const MarginalDistribution &truth) {
-  return standard_deviation(prediction, truth.mean);
-}
-
-struct StandardDeviation : public EvaluationMetric<Eigen::VectorXd> {
-  StandardDeviation() : EvaluationMetric<Eigen::VectorXd>(standard_deviation) {}
-};
-
-/*
- * Computes the negative log likelihood under the assumption that the
- * predictive
- * distribution is multivariate normal.
- */
-static inline double
-negative_log_likelihood(const JointDistribution &prediction,
-                        const MarginalDistribution &truth) {
-  const Eigen::VectorXd mean = prediction.mean - truth.mean;
-  Eigen::MatrixXd covariance(prediction.covariance);
-  if (truth.has_covariance()) {
-    covariance += truth.covariance;
-  }
-  return albatross::negative_log_likelihood(mean, covariance);
-}
-
-static inline double
-negative_log_likelihood(const MarginalDistribution &prediction,
-                        const MarginalDistribution &truth) {
-  const Eigen::VectorXd mean = prediction.mean - truth.mean;
-  Eigen::VectorXd variance(prediction.covariance.diagonal());
-  if (truth.has_covariance()) {
-    variance += truth.covariance.diagonal();
-  }
-  return albatross::negative_log_likelihood(mean, variance.asDiagonal());
-}
-
-template <typename PredictType = JointDistribution>
-struct NegativeLogLikelihood : public EvaluationMetric<PredictType> {
-  NegativeLogLikelihood()
-      : EvaluationMetric<PredictType>(negative_log_likelihood) {}
-};
 
 } // namespace albatross
